requirements.txt
web3>=6.0.0
google-genai
requests
python-dotenv
streamlit>=1.28.0
